pyproject.toml
[project]
name = "llmops-issue-resolver"
version = "0.1.5"
description = """
                Automatically resolves issues of LLMOps projects. 
                Receives as input a repo and an issue to resolve, outputs a repo with the issue resolved.
            """
requires-python = ">=3.9,<3.10"
readme = "README.md"
authors = [
    { name = "Bruno Scaglione", email = "bruno.c.scaglione@gmail.com" }
]
license = {file = "LICENSE"}
dependencies = [
    "black>=24.10.0",
    "langchain-google-genai>=2.0.7",
    "langgraph>=0.2.60",
    "python-dotenv>=1.0.1",
    "typer",
]

[project.urls]
Homepage = "https://github.com/Freelunch-AI/llmops-issue-resolver"
Issues = "https://github.com/Freelunch-AI/llmops-issue-resolver/issues"

[project.scripts]
llmops-issue-resolver-cli = "llmops_issue_resolver.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"
publish-url = "https://upload.pypi.org/legacy/"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
default = true # treated as the lowest priority index 

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E501", "I"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "src/tests/unit_tests",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
    ]
ignore_errors = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
